requirements.txt
agora_token_builder==1.0.0
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
django-anymail==12.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-js-asset==3.0.1
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
kombu==5.4.2
msgpack==1.1.0
packaging==24.2
paypalrestsdk==1.13.3
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
reportlab==4.2.5
requests==2.32.3
service-identity==24.2.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
stripe==11.4.1
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
